+++ examples/async-data/app.js
@@ -53,6 +53,10 @@ class App extends React.Component {
               key={item.abbr}
             >{item.name}</div>
           )}
+          isClearable={true}
+          onClear={ () => this.setState({ value:'' } ) }
+          clearButtonStyle={ { color:'blue' } }
+
         />
       </div>
     )